1. nanf("")를 NAN 매크로로 통일 2. 기타 백업

+++ Includes/Indicators/SimpleMovingAverage.hpp
@@ -19,7 +19,7 @@ class SimpleMovingAverage final : public Indicator {
   size_t sizet_period_;
 
   int count_;
-  long double sum_;
+  double sum_;
   bool can_calculate_;
 
   vector<double> buffer_;
